feat: Generate frames sequentially by default (#10)
* feat: Make frames sequentially by default

Previously, frames were all generated in parallel in independent canvas contexts. This change makes
the default behaviour closer to behaviour in the browser: a shared context is drawn to repeatedly,
allowing frames to preserve traces of their predecessors. Parallel frame generation in separate
contexts is still possible if `pellicola` is passed `renderInParallel` set to `true`.

Closes #5

* docs: Document parallel vs sequential frame rendering